Introduced spell tiers
