+++ api/src/main/java/net/akami/yggdrasil/api/spell/Spell.java
@@ -2,7 +2,13 @@ package net.akami.yggdrasil.api.spell;
 
 import org.spongepowered.api.entity.living.player.Player;
 
+import java.util.List;
+
 public interface Spell {
 
-    void cast(Player caster);
+    List<SpellTier> getTiers();
+
+    default SpellTier getTier(int tier) {
+        return getTiers().get(tier-1);
+    }
 }
