+++ api/src/main/java/net/akami/yggdrasil/api/spell/SpellTier.java
@@ -1,0 +1,8 @@
+package net.akami.yggdrasil.api.spell;
+
+import org.spongepowered.api.entity.living.player.Player;
+
+public interface SpellTier {
+
+    void cast(Player player);
+}
