+++ core/src/main/java/net/akami/yggdrasil/spell/SimpleFireballThrow.java
@@ -1,30 +1,21 @@
 package net.akami.yggdrasil.spell;
 
-import com.flowpowered.math.vector.Vector3d;
 import net.akami.yggdrasil.api.spell.Spell;
-import net.akami.yggdrasil.api.utils.YggdrasilMath;
-import org.spongepowered.api.data.key.Keys;
-import org.spongepowered.api.entity.Entity;
-import org.spongepowered.api.entity.EntityTypes;
-import org.spongepowered.api.entity.living.player.Player;
-import org.spongepowered.api.world.World;
+import net.akami.yggdrasil.api.spell.SpellTier;
 
-import java.util.Optional;
+import java.util.Collections;
+import java.util.List;
 
 public class SimpleFireballThrow implements Spell {
 
+    private List<SpellTier> tiers;
+
+    public SimpleFireballThrow() {
+        this.tiers = Collections.singletonList(new SimpleFireballFirstTier());
+    }
+
     @Override
-    public void cast(Player caster) {
-        Vector3d dir = YggdrasilMath.headRotationToDirection(caster.getHeadRotation());
-        World currentWorld = caster.getWorld();
-        Vector3d ballLocation = caster.getPosition()
-                .add(0, 0.7, 0)
-                .add(dir.mul(3));
-        Entity fireBall = currentWorld.createEntity(EntityTypes.FIREBALL, ballLocation);
-        currentWorld.spawnEntity(fireBall);
-        fireBall.offer(Keys.VELOCITY, dir.mul(1.3));
-        fireBall.offer(Keys.ACCELERATION, dir.mul(0.05));
-        fireBall.offer(Keys.EXPLOSION_RADIUS, Optional.of(2));
-        fireBall.offer(Keys.ATTACK_DAMAGE, 3D);
+    public List<SpellTier> getTiers() {
+        return tiers;
     }
 }
